Fix WPrime spline starts from 0
.. offset needed to be a double not an int to ensure
   we handle recording gaps from powertap etc where
   it is not an whole number of seconds

.. Thanks for Armando Mastracci for highlighting this
   and providing a test file to recreate

Fixes #773

+++ src/WPrime.cpp
@@ -73,7 +73,7 @@ WPrime::setRide(RideFile *input)
     QVector<QPointF> points;
 
     last=0;
-    int offset = 0; // always start from zero seconds (e.g. intervals start at and offset in ride)
+    double offset = 0; // always start from zero seconds (e.g. intervals start at and offset in ride)
     bool first = true;
     if (input->recIntSecs() >= 1) {
         RideFilePoint *lp=NULL;
